+++ cmd/bitcoinbenchmark/main.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/fabiodmferreira/crypto-trading/app"
 	"github.com/fabiodmferreira/crypto-trading/assets"
+	"github.com/fabiodmferreira/crypto-trading/broker"
 	"github.com/fabiodmferreira/crypto-trading/collectors"
 	"github.com/fabiodmferreira/crypto-trading/decisionmaker"
 	"github.com/fabiodmferreira/crypto-trading/trader"
@@ -96,21 +97,6 @@ func (ar *AssetsRepositoryMock) Sell(id primitive.ObjectID, price float32) error
 	return nil
 }
 
-type BrokerMock struct {
-}
-
-func NewBrokerMock() *BrokerMock {
-	return &BrokerMock{}
-}
-
-func (bm *BrokerMock) AddBuyOrder(amount, price float32) error {
-	return nil
-}
-
-func (bm *BrokerMock) AddSellOrder(amount, price float32) error {
-	return nil
-}
-
 func RoundDown(input float64, places int) (newVal float32) {
 	var round float64
 	pow := math.Pow(10, float64(places))
@@ -126,7 +112,7 @@ func benchmark(decisionMakerOptions decisionmaker.DecisionMakerOptions, PriceVar
 	notificationsService := &NotificationsMock{}
 	logService := &LogMock{}
 	account := &AccountMock{5000}
-	broker := NewBrokerMock()
+	broker := broker.NewBrokerMock()
 	assetsRepository := &AssetsRepositoryMock{}
 	trader := trader.NewTrader(assetsRepository, logService, broker)
